Moved note table stuff to its own file
